TexCoord2 support base
