+++ Operators/Lib/Resources/3d/mesh/fx/mesh-UVs.hlsl
@@ -1,11 +1,12 @@
-//#include "lib/shared/hash-functions.hlsl"
-// #include "lib/shared/point.hlsl"
-//#include "lib/shared/quat-functions.hlsl"
-#include "lib/shared/pbr.hlsl"
+//#include "shared/hash-functions.hlsl"
+//#include "shared/point.hlsl"
+//#include "shared/quat-functions.hlsl"
+#include "shared/pbr.hlsl"
 
 cbuffer Params : register(b0)
 {
     float BlendFactor;
+    float SwitchUV;
 }
 
 StructuredBuffer<PbrVertex> VerticesA : t0;        // input
@@ -31,7 +32,13 @@ RWStructuredBuffer<PbrVertex> ResultVertices : u0; // output
 
     float f = BlendFactor;
 
-    ResultVertices[i.x].Position = lerp(A.Position, float3(A.TexCoord,0) , f);
+    if (SwitchUV > 0.5){
+        ResultVertices[i.x].Position = lerp(A.Position, float3(A.TexCoord2,0) , f);
+    }
+    else {
+        ResultVertices[i.x].Position = lerp(A.Position, float3(A.TexCoord,0) , f);
+    }
+
     ResultVertices[i.x].Normal = lerp(A.Normal,float3(0,0,1), f); 
     ResultVertices[i.x].Tangent = lerp(A.Tangent, float3(1,0,0), f);
     ResultVertices[i.x].Bitangent = lerp(A.Bitangent, float3(0,1,0), f);
