+++ Operators/Lib/mesh/_/UVsViewer.t3ui
@@ -19,21 +19,28 @@
         "X": -493.16177,
         "Y": 823.9202
       }
+    },
+    {
+      "InputId": "d9e27d5d-bc00-4482-a4e7-8bc83c577913"/*SwitchUV*/,
+      "Position": {
+        "X": -485.13284,
+        "Y": 1067.2579
+      }
     }
   ],
   "SymbolChildUis": [
     {
       "ChildId": "030d2aed-000e-4582-a9a6-5e14f5eb004f"/*CalcDispatchCount*/,
       "Position": {
-        "X": 298.73434,
-        "Y": 910.3921
+        "X": 276.55276,
+        "Y": 789.40173
       }
     },
     {
       "ChildId": "2a945cb8-4a42-4875-bc8b-a346c3a0720c"/*ComputeShader*/,
       "Position": {
-        "X": 78.32962,
-        "Y": 853.56177
+        "X": 25.900452,
+        "Y": 707.36505
       }
     },
     {
@@ -46,8 +53,8 @@
     {
       "ChildId": "42313d36-3c92-45a3-a29f-66f876c80ed4"/*GetBufferComponents*/,
       "Position": {
-        "X": -151.26566,
-        "Y": 1100.6326
+        "X": -210.82921,
+        "Y": 1041.4742
       }
     },
     {
@@ -79,8 +86,8 @@
         "Y": 44.69636
       },
       "Position": {
-        "X": 298.73434,
-        "Y": 966.3921
+        "X": 288.65182,
+        "Y": 900.85565
       }
     },
     {
@@ -93,8 +100,8 @@
     {
       "ChildId": "d1b12e3f-af1f-4d86-a487-0f8fe3640c9c"/*_MeshBufferComponents*/,
       "Position": {
-        "X": -308.9133,
-        "Y": 1142.6947
+        "X": -347.2269,
+        "Y": 1223.355
       }
     },
     {
@@ -109,6 +116,13 @@
       "Position": {
         "X": -1.2656555,
         "Y": 1057.6326
+      }
+    },
+    {
+      "ChildId": "fad06111-d66e-4d3b-92e4-3f194a6ff790"/*BoolToFloat*/,
+      "Position": {
+        "X": -106.62213,
+        "Y": 956.109
       }
     }
   ],
