+++ Resources/shared/pbr.hlsl
@@ -19,6 +19,7 @@ struct PbrVertex
     float3 Tangent;
     float3 Bitangent;
     float2 TexCoord;
+    float2 TexCoord2;
     float Selected;
     float __padding;
 };
